ref: Split Project serializer into separate index and detail serializers with API namespace
- Created Api::V1::Projects::ProjectIndexSerializer for index action
- Updated Api::V1::Projects::ProjectSerializer for show, update, and delete actions
- Organized serializers under app/serializers/api/v1/projects/
