--- app/serializers/project_serializer.rb
@@ -1,10 +1,0 @@
-class ProjectSerializer < ActiveModel::Serializer
-  attributes :id,
-             :title,
-             :description,
-             :status,
-             :technology_stack,
-             :repository_url,
-             :live_url,
-             :notes
-end
